backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
pyttsx3==2.90
speechrecognition==3.10.0
numpy==1.24.3
typing-extensions==4.8.0